Some project cleanup and refactoring
