--- .editorconfig
@@ -1,9 +1,0 @@
-root = true
-
-[*]
-end_of_line = crlf
-insert_final_newline = true
-charset = utf-8
-indent_size = 2
-indent_style = tab
-tab_width = 2
