--- .eslintrc
@@ -1,23 +1,0 @@
-{
-  "root": true,
-  "env": {
-    "es2022": true,
-    "browser": true,
-    "node": true
-  },
-  "extends": ["eslint:recommended", "plugin:@typescript-eslint/recommended"],
-  "parser": "@typescript-eslint/parser",
-  "parserOptions": {
-    "ecmaVersion": "latest",
-    "sourceType": "module",
-    "ecmaFeatures": {
-      "modules": true
-    }
-  },
-  "plugins": ["@typescript-eslint"],
-  "rules": {
-    "semi": ["error", "always"],
-    "quotes": ["error", "single"],
-    "comma-dangle": ["error", "always-multiline"]
-  }
-}
