--- CHANGELOG.md
@@ -1,48 +1,0 @@
-# CHANGELOG
-
-## 0.0.10
-
-- Add `lerp` function helper to `maths` utils.
-- Add JSDocs for `collisions` utils.
-- Update to latest npm dependencies.
-
-## 0.0.9
-
-- Fix typo on `setDimensions` function for the `Camera` class 🤦.
-- Amend `setDimensions` call in `Game` class.
-
-## 0.0.8
-
-- Add `baseUrl` to `tsconfig.json` to help with resolving import maps.
-- Update to latest npm dependencies.
-
-## 0.0.7
-
-- No longer overrides looping when using `playSound`.
-
-## 0.0.6
-
-- Add favicon to keep browsers happy. 😅
-- Slighty update vscode snippets.
-
-## 0.0.5
-
-- Add support for tile based drawing using the `DrawTile` function.
-
-## 0.0.4
-
-- Add support for flipping of images using the scale argument on the `DrawFrame` and `DrawFrameOrigin` context functions.
-- Update soundhandler `playSound` function.
-
-## 0.0.3
-
-- Added import maps for `engine` and `game` folders.
-- Add a number of vscode snippets.
-
-## 0.0.2
-
-- Some minor changes and typo fixes.
-
-## 0.0.1
-
-- Initial set up and release.
