+++ index.html
@@ -14,7 +14,6 @@
 	</div>
 	<script type="module" src="./src/game/engine/mapGeneration.js"></script>
 	<script type="module" src="./src/game/entities/player_inputs.js"></script>
-	<!-- <script type="module" src="./src/game/colisionMap.js"></script> -->
 </body>
 
 </html>
