--- package.json
@@ -1,22 +1,0 @@
-{
-	"name": "@shezzor/javascript-game-development-template",
-	"type": "module",
-	"version": "0.0.10",
-	"description": "A simple starter template to create JavaScript games",
-	"author": {
-		"name": "Anthony Sherratt",
-		"url": "http://anthonysherratt.co.uk"
-	},
-	"funding": {
-		"type": "patreon",
-		"url": "https://www.patreon.com/shezzor"
-	},
-	"license": "MIT",
-	"main": "./src/index.js",
-	"devDependencies": {
-		"@typescript-eslint/eslint-plugin": "6.3.0",
-		"@typescript-eslint/parser": "6.3.0",
-		"eslint": "8.47.0",
-		"typescript": "5.1.6"
-	}
-}
