+++ src/game/constants/player_infos.js
