--- tsconfig.json
@@ -1,16 +1,0 @@
-{
-	"compilerOptions": {
-		"target": "ESNext",
-		"module": "NodeNext",
-		"lib": ["DOM", "ESNext"],
-		"allowJs": true,
-		"incremental": false,
-		"sourceMap": true,
-		"outDir": "./build",
-		"baseUrl": "./src",
-		"paths": {
-			"engine": ["./engine"],
-			"game": ["./game"]
-		}
-	}
-}
